Refactor and create routes.js. Cut and paste the content of server function in routes. Require fs in routes and create requestHandler function that takes in request and response. Define url and method inside requestHandler. Export requestHandler so it can be accessed by server.js. In server.js, delete everything inside the createServer function and replace it with routes. Require routes by providing file path. Check to make sure everything is working correctly. List alternative ways to export requestHandler.
