+++ message.txt
@@ -1,1 +1,1 @@
-Hello+world%21
+test
