+++ routes.js
@@ -1,0 +1,50 @@
+const fs = require("fs");
+
+const requestHandler = (req, res) => {
+  const url = req.url;
+  const method = req.method;
+  if (url === "/") {
+    res.write("<html>");
+    res.write("<head><title>Enter Message</title></head>");
+    res.write(
+      '<body><form action = "/message" method="POST"><input type = "text" name="message"><button>Send</button></body>'
+    );
+    res.write("</html>");
+    return res.end();
+  }
+  if (url === "/message" && method === "POST") {
+    const body = [];
+    req.on("data", (chunk) => {
+      body.push(chunk);
+    });
+    return req.on("end", () => {
+      const parsedBody = Buffer.concat(body).toString();
+      const message = parsedBody.split("=")[1];
+      fs.writeFileSync("message.txt", message);
+      res.statusCode = 302;
+      res.setHeader("Location", "/");
+      return res.end();
+    });
+  }
+  res.setHeader("Content-Type", "text/html");
+  res.write("<html>");
+  res.write("<head><title>My First Page</title></head>");
+  res.write("<body><h1> Hello from my Node.js Server! </h1></body>");
+  res.write("</html>");
+  res.end();
+};
+
+module.exports = requestHandler;
+
+// alternative ways to write export
+// module.exports = {
+//     handler: requestHandler,
+//     someText: 'Some hard coded text'
+// }
+
+// module.export.handler = requestHandler
+// module.export.someText = "Some hard coded text"
+
+// exports.handler = requestHandler
+// exports.someText = "Some hard coded text"
+
